add support for will_paginate; remove Kaminari dependency
There's no reason why Kaminari runtime dependency needs to be declared,
as api-pagination will operate on any object that responds to:
 - current_page
 - total_pages
 - first_page?
 - last_page?

This monkey-patches WillPaginate::CollectionMethods to make it quack
like Kaminari::PageScopeMethods, so that the plugin supports both
libraries.

A dummy paginated class is provided for the purpose of tests.