Change interface for Rails pagination
This fixes #7 but is also related to a more dire issue. `after_filter`
was misunderstood to provide access to the response before the body had
been written. In actuality, the response has already been fully
constructed by the time `after_filter` is called, which means that the
collection has already been queried and written as json (or XML). In
order to maintain a nice interface without using after_filter, the
method signature of `paginate` has changed to act like `render`:

```rails
class API::v1::PostsController < ApplicationController
  def index
    # Uncomment this line to change the default per_page from 10 to 25:
    # params[:per_page] ||= 25
    posts = Post.all

    paginate json: posts
  end
end
```

At the end, `paginate` simply passes its arguments on to `render` so all
of `render's` options are supported. The collection is assumed to live
either in the argument hash's `:json` key or an `:xml` key.

Signed-off-by: David Celis <me@davidcel.is>