Allow bypassing the implicit render in `paginate`
When using Rails, a common pattern would be to simply paginate a
collection and then render it using a custom serializer. Although this
is currently possible by calling `_paginate_collection` directly, it was
unintended. To better support custom serialization or things like
JBuilder, allow users to bypass the implicit render at the end of
`paginate` by passing their collection to it directly instead of via the
`:json` or `:xml` options:

```ruby
class MoviesController
  def index
    movies = paginate(Movie.all, per_page: 25)

    render json: MoviesSerializer.new(movies)
  end
end
```

If the call to `paginate` still mimics what a call to `render` looks
like, an implicit render will happen. If it looks more like the above,
the collection will simply be paginated and returned after the headers
are set up.

Signed-off-by: David Celis <me@davidcel.is>