Only display paginator warning when relevant
If the user has configured a paginator, the warning about Kaminari and
WillPaginate both appearing in the Gemfile should not be printed.
Additionally, this moves loading of Kaminari and WillPaginate back into
`hooks.rb` with a warning printed if neither gem is included. This
warning still gets printed when the user's application starts. Finally,
add a note to the README saying that we basically aren't going to try to
deal with people's Kaminari/WillPaginate conflicts anymore. Let them
sort that out themselves. Sorry, folks!

Related: #18, #36

Signed-off-by: David Celis <me@davidcel.is>