Add max_per_page value to grape swagger docs
If a max_per_page value is added to grape then it is show in the docs.
This avoids confusing if someone tries to pass in a value greater than
max_per_page as it will return an error so the user can correct
themself. Note this may be a breaking change for some people.